+++ src/main/java/org/jcd2052/api/exceptionhandler/exceptions/DeveloperStudioNotFoundException.java
@@ -1,7 +1,7 @@
 package org.jcd2052.api.exceptionhandler.exceptions;
 
 public class DeveloperStudioNotFoundException extends RuntimeException {
-    public DeveloperStudioNotFoundException(int studioId) {
+    public DeveloperStudioNotFoundException(long studioId) {
         super("Couldn't find studio with id " + studioId);
     }
 }
