+++ src/main/java/org/jcd2052/api/exceptionhandler/exceptions/GameNotFoundException.java
@@ -1,7 +1,7 @@
 package org.jcd2052.api.exceptionhandler.exceptions;
 
 public class GameNotFoundException extends RuntimeException {
-    public GameNotFoundException(int id) {
+    public GameNotFoundException(long id) {
         super(String.format("Couldn't find a game with id %d", id));
     }
 }
